feat: prevent disposable emails for registration
We will be spammed. Also, many users take disposable email addresses to register, which we don't want.

This PR adds a mechanism to prevent users from registering using disposable emails. It's based on a nice package that does the heavy lifting for us.

The PR also adds a new composer command:

```bash
composer monica:update-blacklist-domains
```

which updates the list of forbidden domain names to use.