feat: updated login and register pages (#20)
We slowly need to have our own design system. We start with this.

This is the new login page. The register page looks the same. There are
few pages to change after this, it'll be done in an another PR.

We'll update the right gray part later.

<img width="968" height="778" alt="image"
src="https://github.com/user-attachments/assets/e6afa2f4-4de0-4d2c-8d8e-bdeaeab633c4"
/>

---------

Co-authored-by: Alexis Saettler <alexis@saettler.org>