build: switch package manager from npm to bun (#37)
This PR changes npm to [bun](https://bun.sh/) as the package manager and
bundler.

On average, it runs ~7× faster than npm, ~4× faster than pnpm, and ~17×
faster than yarn.

[This article](https://bun.com/blog/behind-the-scenes-of-bun-install) is
fascinating.

I know this will be a controversial PR. Everything you can do with npm,
you can do with Bun, and Bun is SO. FREAKING. FAST.

Try it out yourself.

If we make the switch, we'll have to update CI probably, and also remove
package-lock.json.

---------

Co-authored-by: Alexis Saettler <alexis@saettler.org>