chore: add laravel boost (#35)
Add [Laravel Boost](https://github.com/laravel/boost), which is the
officiel Laravel support for AI agents as it helps to guide AI agents
and editors.

I've added all the crap that it installs to `.gitignore` so the repo
remains clean.

To install it, just `php artisan boost:install` locally.

---------

Co-authored-by: Alexis Saettler <alexis@saettler.org>