feat: create token login page (#45)
Create a login mechanism for mobile and other devices.

It ressembles to the authorization code flow with PKCE
https://laravel.com/docs/12.x/passport#requesting-auth-pkce-grant-tokens

https://auth0.com/docs/get-started/authentication-and-authorization-flow/authorization-code-flow-with-pkce

It follows some part of https://datatracker.ietf.org/doc/html/rfc7636

To use it, you first need to open a browser to the `/auth/authorize`
url, like this:
```php
    $request->session()->put('state', $state = Str::random(40));

    $request->session()->put(
        'code_verifier', $codeVerifier = Str::random(128)
    );

    $codeChallenge = strtr(rtrim(
        base64_encode(hash('sha256', $codeVerifier, true))
    , '='), '+/', '-_');

    $query = http_build_query([
        'redirect_uri' => 'https://myapp.test/auth/callback',
        'state' => $state,
        'code_challenge' => $codeChallenge,
    ]);

    return redirect('https://rachel.test/auth/authorize?'.$query);
```

After the user logged in and authorized the device, it redirects to the
callback url.

There you can request for a token by sending a POST request to
`/auth/token`:
```php
Route::get('/auth/callback', function (Request $request) {
    $state = $request->session()->pull('state');

    $codeVerifier = $request->session()->pull('code_verifier');

    throw_unless(
        strlen($state) > 0 && $state === $request->state,
        InvalidArgumentException::class
    );

    $response = Http::asForm()
        ->withHeader('User-Agent', $request->header('User-Agent'))
        ->post('https://rachel.test/auth/token', [
            'code_verifier' => $codeVerifier,
            'code' => $request->code,
        ]);
        ->throw();
    
    return $response;
});
```

example response:
```json
{"token_type":"Bearer","access_token":"XPFp1qf1PhEtClGs5B4xGhBe5Lqix0r6ITvOBca8cfbcd841","expires_in":31536000}
```

Note the token is valid for 1 year, but we can change that if needed.

Note: I also added `laravel-debugbar` package.